Move exceptions/logging/rate_limiter into energex.core behind shims
