+++ src/energex/core/logging_config.py
@@ -1,0 +1,53 @@
+"""Centralized logging configuration for energex."""
+
+import logging
+import sys
+from pathlib import Path
+
+
+def setup_logging(
+    log_level: str = "INFO",
+    log_file: Path | str | None = None,
+    enable_console: bool = True,
+) -> None:
+    """
+    Configure application-wide logging.
+
+    Args:
+        log_level: Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL).
+        log_file: Optional file path for log output.
+        enable_console: Whether to log to console. Defaults to True.
+
+    Example:
+        >>> setup_logging(log_level="DEBUG", log_file="energex.log")
+    """
+    handlers: list[logging.Handler] = []
+
+    # Console handler
+    if enable_console:
+        console_handler = logging.StreamHandler(sys.stdout)
+        console_handler.setFormatter(
+            logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+        )
+        handlers.append(console_handler)
+
+    # File handler
+    if log_file:
+        log_path = Path(log_file) if isinstance(log_file, str) else log_file
+        log_path.parent.mkdir(parents=True, exist_ok=True)
+        file_handler = logging.FileHandler(log_path)
+        file_handler.setFormatter(
+            logging.Formatter(
+                "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s:%(lineno)d - %(message)s"
+            )
+        )
+        handlers.append(file_handler)
+
+    # Configure root logger
+    logging.basicConfig(level=getattr(logging, log_level.upper()), handlers=handlers)
+
+    # Silence noisy third-party libraries
+    logging.getLogger("urllib3").setLevel(logging.WARNING)
+    logging.getLogger("yfinance").setLevel(logging.WARNING)
+    logging.getLogger("httpx").setLevel(logging.WARNING)
+    logging.getLogger("httpcore").setLevel(logging.WARNING)
