+++ src/energex/exceptions.py
@@ -1,37 +1,13 @@
-"""Custom exceptions for energex."""
+"""Re-export shim — exceptions moved to energex.core.exceptions (S1).
 
+Kept so the 122 pre-S1 tests and existing modules keep importing energex.exceptions.
+"""
 
-class EnergexError(Exception):
-    """Base exception for all energex errors."""
-
-    pass
-
-
-class ConfigurationError(EnergexError):
-    """Raised when there are configuration issues."""
-
-    pass
-
-
-class LLMProviderError(EnergexError):
-    """Raised when LLM provider operations fail."""
-
-    pass
-
-
-class DataFetchError(EnergexError):
-    """Raised when data fetching operations fail."""
-
-    pass
-
-
-class AnalysisError(EnergexError):
-    """Raised when analysis computations fail."""
-
-    pass
-
-
-class DatabaseError(EnergexError):
-    """Raised when database operations fail."""
-
-    pass
+from energex.core.exceptions import (  # noqa: F401
+    AnalysisError,
+    ConfigurationError,
+    DatabaseError,
+    DataFetchError,
+    EnergexError,
+    LLMProviderError,
+)
